[sites.instagram] rename to instagram_sender
