+++ instagram_to_discord/sites/instagram/instagram_sender.py
@@ -9,7 +9,6 @@ logger.addHandler(handler)
 logger.propagate = False
 
 from typing import List
-import discord
 from .instagram_type import InstagramData
 from ...string_util import sophisticate_string
 from .converter_instagram_url import instagram_make_author_page
