+++ tests/test_twitter.py
@@ -5,7 +5,7 @@ import unittest
 
 from debug import DEBUG
 
-from instagram_to_discord.sites.instagram.instagram import convert_long_caption
+from instagram_to_discord.sites.instagram.instagram_sender import convert_long_caption
 from instagram_to_discord.sites.twitter.twitter_image import convert_twitter
 
 from . import context
